% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa.R
\name{get_chelsa}
\alias{get_chelsa}
\title{Retrieve Monthly CHELSA Data}
\usage{
get_chelsa(where, var = "clt", year = 2023, month = 1)
}
\arguments{
\item{where}{Numeric vector of length 4 specifying a bounding box
in the form \code{c(xmin, ymin, xmax, ymax)} or a matrix/data.frame with
two columns (longitude, latitude) representing points of interest.}

\item{var}{Character string specifying the CHELSA variable to retrieve.
Supported values include:
\describe{
\item{clt}{Total Cloud Cover (monthly average, percentage).}
\item{hurs}{Near-Surface Relative Humidity (monthly average, percentage).}
\item{pr}{Precipitation (monthly total, mm).}
\item{prec}{Precipitation alternative version (mm).}
\item{ps}{Surface Air Pressure (monthly average, Pa).}
\item{rsds}{Surface Downwelling Shortwave Radiation (monthly mean, W / m2).}
\item{sfcWind}{Near-Surface Wind Speed (monthly average, m s^-1).}
\item{tas}{Mean Near-Surface Air Temperature (monthly average, K).}
\item{tasmax}{Maximum Near-Surface Air Temperature (monthly average, K).}
\item{tasmin}{Minimum Near-Surface Air Temperature (monthly average, K).}
\item{tz}{Temperature Lapse Rate (K/m).}
\item{we}{Wind Energy / Wind Power Density (W m^-2).}
}}

\item{year}{Integer specifying the year (1979-2025)}

\item{month}{Integer specifying the month (1-12)}
}
\value{
If \code{where} is a bounding box (vector of length four), a \code{SpatRaster}
object (from the \code{terra} package) containing the requested variable
over the specified area. The raster has time information stored. If
\code{where} is a matrix or data frame, a \code{data.frame} containing
coordinates and corresponding values.
}
\description{
Downloads and returns CHELSA (Climatologies at high resolution for the
earth’s land surface areas) data, a high-resolution global dataset
of climate variables. It features topographically informed downscaling
that accounts for effects such as windward/leeward exposure and
atmospheric boundary layer dynamics. It has a high spatial resolution
of ~1 km (30 arc-sec) and is widely used for species distribution
modeling and ecological research. Data are accessed via the Swiss
Federal Institute for Forest, Snow and Landscape Research (WSL)
and the EnviDat platform.
}
\examples{
\dontrun{
# Bounding box
bbox <- c(44, 25, 63, 40)
tas_raster <- get_chelsa(bbox, var="tas", year=2020, month=6)

# Using coordinates matrix
coords <- cbind(runif(n=10, 44, 63), runif(n=10, 25, 40))
tas_mat <- get_chelsa(coords, var="tas", year=2020, month=6)
}

}
\references{
Karger, D.N., et al. (2017). Climatologies at high resolution for the
earth's land surface areas. Scientific Data, 4, 170122.
\url{https://chelsa-climate.org/}
}
\seealso{
\code{\link[terra]{rast}}
}
\author{
Abdollah Jalilian
}
