% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{get_dem}
\alias{get_dem}
\title{Retrieve Copernicus Digital Elevation Model (DEM) Data}
\usage{
get_dem(where, res = 30, output_dir = tempdir())
}
\arguments{
\item{where}{Either:
\itemize{
\item A numeric vector of length 4 specifying the bounding box
in longitude/latitude format: \code{c(xmin, ymin, xmax, ymax)}.
\item A matrix or data frame with two columns representing
longitude (first column) and latitude (second column) of points.
All coordinates must be in the WGS84 coordinate reference system.
}}

\item{res}{Numeric. Resolution of the DEM data in meters.
Must be either \code{30} or \code{90}.}

\item{output_dir}{Character. Directory to store the downloaded DEM tiles.
Defaults to a temporary directory.}
}
\value{
If \code{where} is a bounding box (numeric vector of length four),
returns a \code{SpatRaster} object representing the merged DEM data.
If \code{where} is a matrix or data frame, returns a \code{data.frame} with
coordinates and their corresponding DEM values.
}
\description{
This function downloads and mosaics Copernicus DEM data from the
Microsoft Planetary Computer. Users can specify a bounding box
or a set of coordinates to retrieve elevation values.
}
\details{
\strong{Important Note:}
Any use of data provided by the Copernicus program must include proper
citation and acknowledgement of the data sources. Users are required to
follow the license and terms of use specified by Copernicus and European
Space Agency. Failure to do so may violate the data usage policies.
}
\examples{
\dontrun{
  # Retrieve DEM for a bounding box
  dem_raster <- get_dem(c(-3, 5, -2, 6), res = 90)
  plot(dem_raster)

  # Retrieve DEM for specific coordinates
  coords <- cbind(runif(n = 100, -3, -2), runif(n = 100, 5, 6))
  dem_points <- get_dem(coords, res = 90)
  print(dem_points)
}

}
\references{
\itemize{
\item Copernicus DEM: European Space Agency (ESA) and the European Commission.
\itemize{
\item Copernicus Digital Elevation Model (Copernicus DEM), distributed by the European Space Agency under the Copernicus Programme.
\item Available at: https://spacedata.copernicus.eu/ and https://doi.org/10.5069/G9028PQB
}
\item Microsoft Planetary Computer: Microsoft AI for Earth.
\itemize{
\item Microsoft Planetary Computer.
\item Available at: https://planetarycomputer.microsoft.com/
}
}
}
\seealso{
\link[terra]{rast}
}
\author{
Abdollah Jalilian
}
