% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_era5.R
\name{get_era5_land}
\alias{get_era5_land}
\title{Download ERA5-Land Climate Data}
\usage{
get_era5_land(
  key,
  vars,
  area,
  year,
  month = NULL,
  day = NULL,
  time = NULL,
  agglevel = "days",
  temp_dir = NULL
)
}
\arguments{
\item{key}{A character string representing the ECMWF API key associated with your user account.}

\item{vars}{A character vector specifying the land surface variables to retrieve. Supported values include:
\itemize{
\item "2m_dewpoint_temperature", "2m_temperature", "skin_temperature"
\item "soil_temperature_level_1", "soil_temperature_level_2", "soil_temperature_level_3", "soil_temperature_level_4"
\item "lake_bottom_temperature", "lake_ice_depth", "lake_ice_temperature", "lake_mix_layer_depth", "lake_mix_layer_temperature"
\item "snow_albedo", "snow_cover", "snow_density", "snow_depth", "snow_depth_water_equivalent", "snowfall", "snowmelt", "temperature_of_snow_layer"
\item "skin_reservoir_content", "volumetric_soil_water_layer_1", "volumetric_soil_water_layer_2", "volumetric_soil_water_layer_3", "volumetric_soil_water_layer_4"
\item "surface_latent_heat_flux", "surface_net_solar_radiation", "surface_net_thermal_radiation", "surface_sensible_heat_flux"
\item "evaporation_from_bare_soil", "evaporation_from_open_water_surfaces_excluding_oceans", "evaporation_from_the_top_of_canopy", "evaporation_from_vegetation_transpiration"
\item "potential_evaporation", "runoff", "snow_evaporation", "sub_surface_runoff", "surface_runoff", "total_evaporation"
\item "10m_u_component_of_wind", "10m_v_component_of_wind", "surface_pressure", "total_precipitation"
\item "leaf_area_index_high_vegetation", "leaf_area_index_low_vegetation", "high_vegetation_cover", "low_vegetation_cover"
\item "lake_total_depth", "land_sea_mask", "soil_type", "type_of_high_vegetation", "type_of_low_vegetation"
}}

\item{area}{A numeric vector of length 4 defining the bounding box for the region of interest,
in the format \code{c(North, West, South, East)}. Coordinates must fall within:
- North: 90.0째 N
- South: -90.0째 N
- West: -180.0째 W
- East: 180.0째 W}

\item{year}{A numeric value specifying the year to retrieve data for (from 1950 onwards).}

\item{month}{A numeric or character vector specifying the months to retrieve (default: all months in the year).
Values are automatically formatted as two-digit strings (e.g., \code{"01"}, \code{"02"}).}

\item{day}{A numeric or character vector specifying the days to retrieve (default: all days for the specified months).
Values are automatically formatted as two-digit strings (e.g., \code{"01"}, \code{"02"}).}

\item{time}{A character vector specifying the hours to retrieve (default: all 24 hours of the day).
The values are automatically formatted as two-digit hour strings (e.g., \code{"00:00"}, \code{"01:00"}).}

\item{agglevel}{A character string specifying the temporal aggregation level.
Options are:
\itemize{
\item \code{NULL} (no aggregation)
\item \code{"years"}
\item \code{"months"}
\item \code{"yearmonths"}
\item \code{"dekads"} (10-day periods)
\item \code{"yeardekads"}
\item \code{"weeks"} (ISO 8601 week number)
\item \code{"yearweeks"}
\item \code{"days"} (default)
\item \code{"doy"} (day of the year)
\item \code{"7days"} (seven-day periods starting at Jan 1 of each year)
\item \code{"10days"}
\item \code{"15days"}
}}

\item{temp_dir}{A character string specifying a temporary directory for downloaded
and extracted files. If \code{NULL} (default), the system's temporary directory
(\code{tempdir()}) is used.}
}
\value{
A \code{SpatRaster} object containing the downloaded air quality data.
}
\description{
This function retrieves ERA5-Land data from the Copernicus Data Store via the ECMWF API.
ERA5-Land provides a detailed, long-term record of land surface variables from 1950 to the present.
It is derived from the ERA5 climate reanalysis model, using both model data and observations,
and adjusted for elevation.
}
\details{
This function retrieves data from the ERA5-Land dataset, which covers the period from January 1950 to
approximately 2-3 months before the present. The data is processed into a \code{SpatRaster} object for spatial analysis.
Temporal aggregation (e.g., by months or years) can be performed after retrieval if needed.
}
\examples{
\dontrun{
# Define the geographical area (North, West, South, East)
area <- c(47.1, 6.6, 35.4, 18.6)

# Set ECMWF API key (replace with your actual key)
key <- "********************************"

# Download skin temperature data for 6th of October 2022
st <- get_era5_land(key, vars = "skin_temperature", area = area,
                    year = 2022, month = 10, day=6)

# Plot the retrieved data
plot(st)
}

}
\references{
Copernicus Climate Data Store: https://cds.climate.copernicus.eu/
ECMWF API Documentation: https://confluence.ecmwf.int/display/CKB/How+to+download+ERA5
}
\author{
Abdollah Jalilian
}
