% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lulc.R
\name{get_lulc}
\alias{get_lulc}
\title{Retrieve 10m Annual Land Use Land Cover from 2017 to 2023}
\usage{
get_lulc(
  where,
  year = 2023,
  labels = FALSE,
  fact = 0,
  output_dir = tempdir(),
  clean_dir = FALSE
)
}
\arguments{
\item{where}{A numeric vector of length 4 representing the bounding box in the
form c(xmin, ymin, xmax, ymax) or a matrix/data.frame with two columns
(longitude, latitude) representing points of interest.}

\item{year}{A numeric value specifying the year of interest (default is 2023).
Valid values are between 2017 and 2023.}

\item{labels}{A logical value indicating whether to return land use and land
cover class names (default is FALSE). If TRUE, the function returns
descriptive labels for each class.}

\item{fact}{A numeric factor for aggregation. If greater than 0, the output raster
will be aggregated by the specified factor using the modal function.}

\item{output_dir}{A character string specifying the output directory for storing
downloaded data (default is a temporary directory).}

\item{clean_dir}{A logical value indicating whether to clean up the output
directory after the function has completed (default is FALSE).}
}
\value{
A raster or data frame containing the land use and land cover data for
the specified area and year. If \code{where} is a matrix or data frame, the
function returns a data frame with coordinates and corresponding LULC values.
}
\description{
This function retrieves global land use and land cover (LULC) maps derived from
ESA Sentinel-2 imagery at a 10m resolution for the years 2017-2023. The dataset
includes 9 LULC classes and is created using a deep learning model trained on
billions of human-labeled pixels curated by the National Geographic Society.
Please follow the license and terms of use from Impact Observatory (processor,
producer, licensor), Microsoft (host), and Esri (licensor). Failure to comply
may result in data usage policy violations.
}
\details{
The function first validates the input, including the \code{where} parameter
(bounding box or coordinates) and the \code{year} parameter. It then uses the Microsoft
Planetary Computer's STAC API to search and download LULC data tiles, which are
cropped to the specified bounding box. The function returns a raster object with
merged LULC data or a data frame with LULC values for the specified coordinates.

The LULC classes are represented by integer values:
\itemize{
\item 0: No Data
\item 1: Water
\item 2: Trees
\item 4: Flooded vegetation
\item 5: Crops
\item 7: Built area
\item 8: Bare ground
\item 9: Snow/ice
\item 10: Clouds
\item 11: Rangeland
}

\strong{Important Note:}
This dataset is produced by Impact Observatory (processor, producer, licensor),
Microsoft (host), and Esri (licensor). Users are required to follow the
license and terms of use specified by Impact Observatory and Esri.
Failure to do so may violate the data usage policies.
}
\examples{
\dontrun{
# Retrieve LULC for a bounding box in 2023 and include class labels
lulc_data <- get_lulc(c(47, 34, 47.5, 35), year = 2023, labels = TRUE)

# Retrieve LULC for a set of coordinates with aggregation
coordinates <- data.frame(lon = runif(50, 47, 47.5), lat=runif(50, 34, 35))
lulc_data <- get_lulc(coordinates, year = 2023, fact = 2)
}

}
\references{
-Impact Observatory, Microsoft, and Esri. (2023). Global Land Use Land Cover (LULC) Dataset, 10m Resolution (2017-2023).
\itemize{
\item ESA Sentinel-2 Imagery. Available at: https://planetarycomputer.microsoft.com/
}
}
\seealso{
\link[terra]{rast}, \link[terra]{focal}
}
\author{
Abdollah Jalilian
}
