% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modis.R
\name{get_modis}
\alias{get_modis}
\title{Retrive MODIS Data}
\usage{
get_modis(
  where,
  var,
  datetime,
  collection = NULL,
  crop = TRUE,
  w = NA,
  agglevel = NULL,
  output_dir = tempdir(),
  clean_dir = FALSE
)
}
\arguments{
\item{where}{A numeric vector of length 4 representing the bounding box in the
form c(xmin, ymin, xmax, ymax) or a matrix/data.frame with two columns
(longitude, latitude) representing points of interest.}

\item{var}{A character string specifying the MODIS variable to be retrieved.
see the details for more explanation.}

\item{datetime}{A character string specifying the date range in the format
"YYYY-MM-DD/YYYY-MM-DD".}

\item{collection}{A character string specifying the name of the MODIS collection
from which the variable should be retrieved. If \code{NULL} (default), the
function will automatically determine the appropriate collection based
on the selected variable.}

\item{crop}{Logical, whether to crop the retrieved raster data to the bounding box. Default is TRUE.}

\item{w}{Window size for the moving window method to fill missing (NA) values using the \code{focal} function from the \code{terra} package.
If \code{w = NA} (default), missing values are not filled. See the documentation of \code{terra::focal} for more details.}

\item{agglevel}{A character string specifying the temporal aggregation level for the retrieved data, using the \code{aggregate} function
in the \code{terra} package. Available options include:
- \code{NULL} (no aggregation)
- \code{"years"}, \code{"months"}, \code{"yearmonths"}
- \code{"dekads"} (10-day periods), \code{"yeardekads"}
- \code{"weeks"} (ISO 8601 week number), \code{"yearweeks"}
- \code{"days"} (default), \code{"doy"} (day of the year)
- \code{"7days"}, \code{"10days"}, \code{"15days"}}

\item{output_dir}{A character string specifying the directory where downloaded files should be saved.
Default is a temporary directory.}

\item{clean_dir}{Logical, whether to clean the output directory before downloading new files. Default is FALSE.}
}
\description{
Access and download the Moderate Resolution Imaging Spectroradiometer (MODIS)
satellite data, available through the Microsoft's Planetary Computer STAC API.
It uses the STAC API, processes the retrieved raster data, and optionally
extracts values for given spatial points.
}
\details{
This function allows users to retrieve MODIS satellite data for specified
geographic regions and time periods. MODIS provides various environmental
datasets, such as land surface temperature, vegetation indices, burned areas,
and primary productivity. The following table provides information on available
MODIS collections and their corresponding variables:

\tabular{lll}{
\strong{id} \tab \strong{title} \tab \strong{vars} \cr
modis-64A1-061 \tab MODIS Burned Area Monthly \tab QA, hdf, Last_Day, metadata, Burn_Date, First_Day, Burn_Date_Uncertainty \cr
modis-17A2H-061 \tab MODIS Gross Primary Productivity 8-Day \tab hdf, Gpp_500m, metadata, PsnNet_500m, Psn_QC_500m \cr
modis-11A2-061 \tab MODIS Land Surface Temperature/Emissivity 8-Day \tab hdf, QC_Day, Emis_31, Emis_32, QC_Night, metadata, LST_Day_1km, Day_view_angl, Day_view_time, LST_Night_1km, Clear_sky_days, Night_view_angl, Night_view_time, Clear_sky_nights \cr
modis-17A2HGF-061 \tab MODIS Gross Primary Productivity 8-Day Gap-Filled \tab hdf, Gpp_500m, metadata, PsnNet_500m, Psn_QC_500m \cr
modis-17A3HGF-061 \tab MODIS Net Primary Production Yearly Gap-Filled \tab hdf, Gpp_500m, Npp_500m, metadata, Npp_QC_500m \cr
modis-09A1-061 \tab MODIS Surface Reflectance 8-Day (500m) \tab hdf, metadata, sur_refl_b01, sur_refl_b02, sur_refl_b03, sur_refl_b04, sur_refl_b05, sur_refl_b06, sur_refl_b07, sur_refl_raz, sur_refl_szen, sur_refl_vzen, sur_refl_qc_500m, sur_refl_state_500m, sur_refl_day_of_year \cr
modis-16A3GF-061 \tab MODIS Net Evapotranspiration Yearly Gap-Filled \tab hdf, ET_500m, LE_500m, PET_500m, PLE_500m, metadata, ET_QC_500m \cr
modis-21A2-061 \tab MODIS Land Surface Temperature/3-Band Emissivity 8-Day \tab hdf, QC_Day, Emis_29, Emis_31, Emis_32, QC_Night, metadata, LST_Day_1KM, LST_Night_1KM, View_Time_Day, View_Angle_Day, View_Time_Night, View_Angle_Night \cr
modis-43A4-061 \tab MODIS Nadir BRDF-Adjusted Reflectance (NBAR) Daily \tab hdf, metadata, Nadir_Reflectance_Band1, Nadir_Reflectance_Band2, Nadir_Reflectance_Band3, Nadir_Reflectance_Band4, Nadir_Reflectance_Band5, Nadir_Reflectance_Band6, Nadir_Reflectance_Band7, BRDF_Albedo_Band_Mandatory_Quality_Band1, BRDF_Albedo_Band_Mandatory_Quality_Band2, BRDF_Albedo_Band_Mandatory_Quality_Band3, BRDF_Albedo_Band_Mandatory_Quality_Band4, BRDF_Albedo_Band_Mandatory_Quality_Band5, BRDF_Albedo_Band_Mandatory_Quality_Band6, BRDF_Albedo_Band_Mandatory_Quality_Band7 \cr
modis-09Q1-061 \tab MODIS Surface Reflectance 8-Day (250m) \tab hdf, metadata, sur_refl_b01, sur_refl_b02, sur_refl_qc_250m, sur_refl_state_250m \cr
modis-14A1-061 \tab MODIS Thermal Anomalies/Fire Daily \tab QA, hdf, MaxFRP, sample, FireMask, metadata \cr
modis-13Q1-061 \tab MODIS Vegetation Indices 16-Day (250m) \tab hdf, metadata, 250m_16_days_EVI, 250m_16_days_NDVI, 250m_16_days_VI_Quality, 250m_16_days_MIR_reflectance, 250m_16_days_NIR_reflectance, 250m_16_days_red_reflectance, 250m_16_days_blue_reflectance, 250m_16_days_sun_zenith_angle, 250m_16_days_pixel_reliability, 250m_16_days_view_zenith_angle, 250m_16_days_relative_azimuth_angle, 250m_16_days_composite_day_of_the_year \cr
modis-14A2-061 \tab MODIS Thermal Anomalies/Fire 8-Day \tab QA, hdf, FireMask, metadata \cr
modis-15A2H-061 \tab MODIS Leaf Area Index/FPAR 8-Day \tab hdf, Lai_500m, metadata, Fpar_500m, FparLai_QC, FparExtra_QC, LaiStdDev_500m, FparStdDev_500m \cr
modis-11A1-061 \tab MODIS Land Surface Temperature/Emissivity Daily \tab hdf, QC_Day, Emis_31, Emis_32, QC_Night, metadata, LST_Day_1km, Clear_day_cov, Day_view_angl, Day_view_time, LST_Night_1km, Clear_night_cov, Night_view_angl, Night_view_time \cr
modis-15A3H-061 \tab MODIS Leaf Area Index/FPAR 4-Day \tab hdf, Lai_500m, metadata, Fpar_500m, FparLai_QC, FparExtra_QC, LaiStdDev_500m, FparStdDev_500m \cr
modis-13A1-061 \tab MODIS Vegetation Indices 16-Day (500m) \tab hdf, metadata, 500m_16_days_EVI, 500m_16_days_NDVI, 500m_16_days_VI_Quality, 500m_16_days_MIR_reflectance, 500m_16_days_NIR_reflectance, 500m_16_days_red_reflectance, 500m_16_days_blue_reflectance, 500m_16_days_sun_zenith_angle, 500m_16_days_pixel_reliability, 500m_16_days_view_zenith_angle, 500m_16_days_relative_azimuth_angle, 500m_16_days_composite_day_of_the_year \cr
modis-10A2-061 \tab MODIS Snow Cover 8-day \tab hdf, metadata, Maximum_Snow_Extent, Eight_Day_Snow_Cover \cr
modis-10A1-061 \tab MODIS Snow Cover Daily \tab hdf, NDSI, metadata, orbit_pnt, granule_pnt, NDSI_Snow_Cover, Snow_Albedo_Daily_Tile, NDSI_Snow_Cover_Basic_QA, NDSI_Snow_Cover_Algorithm_Flags_QA
}

MODIS data is available for different time intervals (daily, 8-day, 16-day,
or yearly) and is provided in 10° x 10° tiles at the equator.
}
\examples{
\dontrun{
  # Retrieve 8-day daytime 1km grid land surface temperature for a bounding box
  temp <- get_modis(c(-3, 5, -2, 6), var = "LST_Day_1KM",
                    datetime = "2023-11-01/2024-02-28")
  plot(temp)

  # Retrieve yearly total evapotranspiration for specific coordinates
  coords <- cbind(runif(n = 100, -3, -2), runif(n = 100, 5, 6))
  et_points <- get_modis(coords, var="ET_500m",
                        datetime = "2023-11-01/2024-02-28")
  print(et_points)
}

}
\seealso{
\link[terra]{aggregate}, \link[terra]{focal}, \link[getsat]{check_modis}
}
\author{
Abdollah Jalilian
}
