% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_popdensity.R
\name{get_popdensity}
\alias{get_popdensity}
\title{Retrieve Gridded Population of the World, Version 4:}
\usage{
get_popdensity(where, year = 2020, res = "30_sec", downloaddir = NULL)
}
\arguments{
\item{where}{A numeric vector of length 4 (bounding box) or a matrix/data.frame with two columns (longitude, latitude) specifying the spatial region of interest. If a bounding box is provided, the coordinates should be in the format: c(xmin, ymin, xmax, ymax).}

\item{year}{An integer specifying the year for which the population density data is required. Valid values are: 2000, 2005, 2011, 2015, and 2020. The default is 2020.}

\item{res}{A character string specifying the resolution of the population density data. Options include:
\itemize{
\item "30_sec" for 30 arc-seconds (~1 km at the equator),
\item "2pt5_min" for 2.5 arc-minutes,
\item "15_min" for 15 arc-minutes,
\item "30_min" for 30 arc-minutes,
\item "1_deg" for 1 degree resolution.
The default is "30_sec".
}}

\item{downloaddir}{A character string specifying the directory where the data file should be downloaded. If not specified, the system's default download directory is used (Windows: \code{file.path(Sys.getenv("USERPROFILE"), "Downloads")}, macOS/Linux: \code{file.path(Sys.getenv("HOME"), "Downloads")}).}
}
\value{
If the input \code{where} is a bounding box, the function returns a raster
object cropped to that region. If \code{where} is a set of coordinates
(matrix/data.frame), the function returns the population density values at
those specific locations.
}
\description{
Access and download the Population Density Adjusted to Match 2015 Revision
UN WPP Country Totals, Revision 11, provided by the Socioeconomic Data and
Applications Center (SEDAC) at NASA. The function returns population density
data adjusted to match the 2015 Revision of the United Nations World Population
Prospects (UN WPP) country totals. The dataset represents estimates of human
population density (people per square kilometer) for various years, using a
proportional allocation gridding algorithm that assigns UN WPP-adjusted
population counts to global grid cells.
}
\details{
The available years are 2000, 2005, 2011, 2015, and 2020.

The data are provided as global rasters at the following resolutions:
\itemize{
\item 30 arc-second (~1 km resolution at the equator),
\item 2.5 arc-minute (~5 km resolution at the equator),
\item 15 arc-minute (~30 km resolution at the equator),
\item 30 arc-minute (~55 km resolution at the equator), and
\item 1 degree (~110 km resolution at the equator),
}

These population density estimates are based on national and sub-national
population counts and spatial distribution information, adjusted to match t
he UN WPP country totals. The rasters are available for download and can be
cropped to a specified bounding box or set of coordinates.

For more information on the dataset and access to the data, visit:
\url{https://www.earthdata.nasa.gov/data/projects/gpw}
}
\examples{
\dontrun{
 # Get population density for a bounding box
 pd1 <- get_popdensity(c(-5, 4, 1, 10), year=2020, res="1_deg")
 terra::plot(pd1)

 # Get population density for specific coordinates
 coords <- cbind(runif(n = 100, -5, -1), runif(n = 100, 4, 10))
 pd2 <- get_popdensity(coords, year=2020, res="1_deg")
 print(pd2)
}

}
\references{
\itemize{
\item Center for International Earth Science Information Network-CIESIN-Columbia University. (2018).
Gridded Population of the World, Version 4 (GPWv4): Population Density
Adjusted to Match 2015 Revision UN WPP Country Totals, Revision 11
(Version 4.11) Data set. Palisades, NY: NASA Socioeconomic Data and
Applications Center (SEDAC). \url{https://doi.org/10.7927/H4F47M65}
\item NASA Earth Data: \url{https://www.earthdata.nasa.gov/data/projects/gpw}
}
}
\author{
Abdollah Jalilian
}
