% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa.R
\name{get_chelsa}
\alias{get_chelsa}
\title{Retrieve Monthly CHELSA Data}
\usage{
get_chelsa(where, var = "clt", year = 2023, month = 1)
}
\arguments{
\item{where}{Numeric vector of length 4 specifying a bounding box
in the form \code{c(xmin, ymin, xmax, ymax)}
or a matrix/data.frame with two columns (longitude, latitude)
representing points of interest.}

\item{var}{Character string specifying the CHELSA variable to retrieve.
Supported values include:
\describe{
\item{\code{"clt"}}{Total Cloud Cover (monthly average; \\%), available for both daily and monthly datasets}
\item{\code{"hurs"}}{Near-Surface Relative Humidity (monthly average; \\%), both daily and monthly}
\item{\code{"pr"}}{Precipitation (monthly total; mm), both daily and monthly}
\item{\code{"prec"}}{Precipitation (alternative version; mm), both daily and monthly}
\item{\code{"ps"}}{Surface Air Pressure (monthly average; Pa), both daily and monthly}
\item{\code{"rsds"}}{Surface Downwelling Shortwave Radiation (monthly mean; W m^-2), both daily and monthly}
\item{\code{"sfcWind"}}{Near-Surface Wind Speed (monthly average; m s^-1), both daily and monthly}
\item{\code{"tas"}}{Mean Near-Surface Air Temperature (monthly average; K), both daily and monthly}
\item{\code{"tasmax"}}{Maximum Near-Surface Air Temperature (monthly average; K), both daily and monthly}
\item{\code{"tasmin"}}{Minimum Near-Surface Air Temperature (monthly average; K), both daily and monthly}
\item{\code{"tz"}}{Temperature Lapse Rate (K m^-1), both daily and monthly}
\item{\code{"we"}}{Wind Energy / Wind Power Density (daily specific; W m^-2), daily-only (year >= 2022)}
}}

\item{year}{Integer (1979-2025).}

\item{month}{Integer (1-12).}
}
\value{
If \code{where} is a bounding box (vector of length four), a \code{SpatRaster}
object (from the \code{terra} package) containing the requested variable
over the specified area. The raster has time information stored. If
\code{where} is a matrix or data frame, a \code{data.frame} containing
coordinates and corresponding values.
}
\description{
Downloads and returns CHELSA (Climatologies at high resolution for the
earth’s land surface areas) data, a high-resolution global dataset
of climate variables. It features topographically informed downscaling
that accounts for effects such as windward/leeward exposure and
atmospheric boundary layer dynamics. It has a high spatial resolution
of ~1 km (30 arc-sec) and is widely used for species distribution
modeling and ecological research. Data are accessed via the Swiss
Federal Institute for Forest, Snow and Landscape Research (WSL)
and the EnviDat platform.
}
\examples{
\dontrun{
# Bounding box
bbox <- c(44, 25, 63, 40)
tas_raster <- get_chelsa(bbox, var="tas", year=2020, month=6)

# Using coordinates matrix
coords <- cbind(runif(n=10, 44, 63), runif(n=10, 25, 40))
tas_mat <- get_chelsa(coords, var="tas", year=2020, month=6)
}

}
\references{
\itemize{
\item Karger, D.N., Conrad, O., Bohner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W.,
Zimmermann, N.E., Linder, H.P., Kessler, M., 2017.
Climatologies at high resolution for the earth’s land surface areas.
Scientific Data, 4, 170122.
Available at: \url{https://chelsa-climate.org/}
\item Karger, D.N., Conrad, O., Bohner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W.,
Zimmermann, N.E., Linder, P., Kessler, M., 2021.
CHELSA v2.1: Climatologies at high resolution for the earth’s land surface areas.
EnviDat. \doi{10.16904/envidat.228}
}
}
\seealso{
\link[terra]{rast}
}
\author{
Abdollah Jalilian
}
