% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem2.R
\name{get_dem2}
\alias{get_dem2}
\title{Retrieve GMTED2010 Digital Elevation Model (DEM) Data}
\usage{
get_dem2(where, res = 0.0625, output_dir = tempdir())
}
\arguments{
\item{where}{Either:
\itemize{
\item A numeric vector of length 4 specifying the bounding box
in longitude/latitude format: \code{c(xmin, ymin, xmax, ymax)}.
\item A matrix or data frame with two columns representing
longitude (first column) and latitude (second column) of points.
All coordinates must be in the WGS84 coordinate reference system.
}}

\item{res}{Numeric. Resolution of the DEM data in degrees.
Must be either \code{0.0625}, \code{0.125}, \code{0.250}, \code{0.500},
\code{0.750} or \code{1.000}.}

\item{output_dir}{Character. Directory to store the downloaded DEM tiles.
Defaults to a temporary directory.}
}
\value{
If \code{where} is a bounding box (numeric vector of length four),
returns a \code{SpatRaster} object representing the merged DEM data.
If \code{where} is a matrix or data frame, returns a \code{data.frame} with
coordinates and their corresponding DEM values.
}
\description{
This function downloads and extracts elevation data from the Global
Multi-resolution Terrain Elevation Data 2010 (GMTED2010) dataset as provided
by the Tropospheric Emission Monitoring Internet Service (TEMIS;
\url{https://www.temis.nl}).
}
\details{
GMTED2010 is a global digital elevation model jointly developed by the
U.S. Geological Survey (USGS) and the National Geospatial-Intelligence Agency
(NGA). The version hosted by TEMIS is a post-processed and resampled variant
of the original USGS/NGA data, provided at several reduced spatial resolutions.

The GMTED2010 dataset is in the public domain and available for use without
restriction.
}
\examples{
\dontrun{
  # Retrieve DEM for a bounding box
  dem_raster <- get_dem2(c(6, 35, 19, 47), res=0.5)
  plot(dem_raster)

  # Retrieve DEM for specific coordinates
  coords <- cbind(runif(n=100, 6, 19), runif(n=100, 35, 47))
  dem_points <- get_dem2(coords, res=0.5)
  print(dem_points)
}

}
\references{
\itemize{
\item Danielson, J.J., and Gesch, D.B., 2011, Global multi-resolution terrain
elevation data 2010 (GMTED2010): U.S. Geological Survey Open-File Report
2011–1073, 26 p. http://pubs.usgs.gov/of/2011/1073/
\item TEMIS GMTED2010 data: https://www.temis.nl/data/gmted2010/
}
}
\seealso{
\link[terra]{rast}
}
\author{
Abdollah Jalilian
}
